Split nadeko into a plugin crate and a runtime crate
This gets all the tests to pass. I'm a little concerned
though that I had to add macro support for matches,
loops, and breaks, which I worry could slip in some
non-constant time operations. Instead it might be
better to pull the `for` loop out of the ChaCha20
test
