+++ tests/chacha20.rs
@@ -1,14 +1,14 @@
 // this code is based from suruga: crypto/chacha20.rs
 
-#![feature(phase, slicing_syntax, macro_rules)]
+#![feature(plugin)]
+#![plugin(nadeko_plugin)]
 
-#[phase(plugin, link)]
 extern crate nadeko;
 
 // convert $e.slice($i, $i + 4) into u32
 macro_rules! to_le_u32(
     ($e:ident[$i:expr]) => ({
-        let i: uint = $i;
+        let i: usize = $i;
         let v1 = $e[i + 0] as u32;
         let v2 = $e[i + 1] as u32;
         let v3 = $e[i + 2] as u32;
@@ -20,7 +20,7 @@ macro_rules! to_le_u32(
 #[allow(missing_copy_implementations)]
 pub struct ChaCha20 {
     // SECRET
-    vals: [u32, ..16],
+    vals: [u32; 16],
 }
 
 impl ChaCha20 {
@@ -29,7 +29,7 @@ impl ChaCha20 {
         assert_eq!(key.len(), 32);
         assert_eq!(nonce.len(), 8);
 
-        let mut vals = [0u32, ..16];
+        let mut vals = [0u32; 16];
 
         // "expand 32-byte k"
         vals[0] = 0x61707865;
@@ -37,7 +37,7 @@ impl ChaCha20 {
         vals[2] = 0x79622d32;
         vals[3] = 0x6b206574;
 
-        for i in range(0u, 8) {
+        for i in 0 .. 8 {
             vals[4 + i] = to_le_u32!(key[4 * i]);
         }
 
@@ -53,15 +53,15 @@ impl ChaCha20 {
         }
     }
 
-    fn round20(&self) -> [u32, ..16] {
+    fn round20(&self) -> [u32; 16] {
         #[const_time]
-        fn round20_inner(vals: &mut [u32, ..16]) {
-            for _ in 0u..10 {
+        fn round20_inner(vals: &mut [u32; 16]) {
+            for _ in 0usize .. 10 {
                 // $e must be > 0 and < 32
                 macro_rules! rot(
                     ($a:expr, $e:expr) => ({
                         let a: u32 = $a;
-                        let e: uint = $e;
+                        let e: usize = $e;
                         (a << e) | (a >> (32 - e))
                     })
                 );
@@ -109,14 +109,14 @@ impl ChaCha20 {
         let mut vals = self.vals;
         round20_inner(&mut vals);
 
-        for i in range(0u, 16) {
-            vals[i] += self.vals[i];
+        for i in 0 .. 16 {
+            vals[i] = vals[i].wrapping_add(self.vals[i]);
         }
 
         vals
     }
 
-    pub fn next(&mut self) -> [u8, ..64] {
+    pub fn next(&mut self) -> [u8; 64] {
         let next = self.round20();
 
         // in TLS, vals[13] never increases
@@ -129,8 +129,8 @@ impl ChaCha20 {
         }
 
         let next_bytes = {
-            let mut next_bytes = [0u8, ..64];
-            for i in range(0u, 16) {
+            let mut next_bytes = [0u8; 64];
+            for i in 0 .. 16 {
                 next_bytes[4 * i + 0] = next[i] as u8;
                 next_bytes[4 * i + 1] = (next[i] >> 8) as u8;
                 next_bytes[4 * i + 2] = (next[i] >> 16) as u8;
@@ -167,29 +167,29 @@ mod test {
 
     fn check_keystream(key: &[u8], nonce: &[u8], keystream: &[u8]) {
         let mut chacha = ChaCha20::new(key, nonce);
-        let input = Vec::from_elem(keystream.len(), 0u8);
-        let output = chacha.encrypt(input[]);
-        assert_eq!(output[], keystream);
+        let input = vec![0u8; keystream.len()];
+        let output = chacha.encrypt(&input);
+        assert_eq!(output, keystream);
     }
 
     #[test]
     fn test_chacha20() {
         // from https://tools.ietf.org/html/draft-agl-tls-chacha20poly1305-04
 
-        let mut key = Vec::from_elem(32, 0u8);
-        let mut nonce = Vec::from_elem(8, 0u8);
+        let mut key = vec![0u8; 32];
+        let mut nonce = vec![0u8; 8];
         let keystream = b"\x76\xb8\xe0\xad\xa0\xf1\x3d\x90\x40\x5d\x6a\xe5\x53\x86\xbd\x28\
                           \xbd\xd2\x19\xb8\xa0\x8d\xed\x1a\xa8\x36\xef\xcc\x8b\x77\x0d\xc7\
                           \xda\x41\x59\x7c\x51\x57\x48\x8d\x77\x24\xe0\x3f\xb8\xd8\x4a\x37\
                           \x6a\x43\xb8\xf4\x15\x18\xa1\x1c\xc3\x87\xb6\x69\xb2\xee\x65\x86";
-        check_keystream(key[], nonce[], keystream);
+        check_keystream(&key, &nonce, keystream);
 
         key[31] = 1;
         let keystream = b"\x45\x40\xf0\x5a\x9f\x1f\xb2\x96\xd7\x73\x6e\x7b\x20\x8e\x3c\x96\
                           \xeb\x4f\xe1\x83\x46\x88\xd2\x60\x4f\x45\x09\x52\xed\x43\x2d\x41\
                           \xbb\xe2\xa0\xb6\xea\x75\x66\xd2\xa5\xd1\xe7\xe2\x0d\x42\xaf\x2c\
                           \x53\xd7\x92\xb1\xc4\x3f\xea\x81\x7e\x9a\xd2\x75\xae\x54\x69\x63";
-        check_keystream(key[], nonce[], keystream);
+        check_keystream(&key, &nonce, keystream);
 
         key[31] = 0;
         nonce[7] = 1;
@@ -197,7 +197,7 @@ mod test {
                           \x0b\x49\xe0\x15\xad\xbf\xf7\x13\x4f\xcb\x7d\xf1\x37\x82\x10\x31\
                           \xe8\x5a\x05\x02\x78\xa7\x08\x45\x27\x21\x4f\x73\xef\xc7\xfa\x5b\
                           \x52\x77\x06\x2e\xb7\xa0\x43\x3e\x44\x5f\x41\xe3";
-        check_keystream(key[], nonce[], keystream);
+        check_keystream(&key, &nonce, keystream);
 
         key[31] = 0;
         nonce[7] = 0;
@@ -206,12 +206,12 @@ mod test {
                           \x09\x63\x16\x34\xd2\x1e\x42\xac\x33\x96\x0b\xd1\x38\xe5\x0d\x32\
                           \x11\x1e\x4c\xaf\x23\x7e\xe5\x3c\xa8\xad\x64\x26\x19\x4a\x88\x54\
                           \x5d\xdc\x49\x7a\x0b\x46\x6e\x7d\x6b\xbd\xb0\x04\x1b\x2f\x58\x6b";
-        check_keystream(key[], nonce[], keystream);
+        check_keystream(&key, &nonce, keystream);
 
-        for i in range(0u, 0x20) {
+        for i in 0 .. 0x20 {
             key[i] = i as u8;
         }
-        for i in range(0u, 0x08) {
+        for i in 0 .. 0x08 {
             nonce[i] = i as u8;
         }
         let keystream = b"\xf7\x98\xa1\x89\xf1\x95\xe6\x69\x82\x10\x5f\xfb\x64\x0b\xb7\x75\
@@ -230,7 +230,7 @@ mod test {
                           \x1c\x89\x4c\x94\xa3\x71\x87\x6a\x94\xdf\x76\x28\xfe\x4e\xaa\xf2\
                           \xcc\xb2\x7d\x5a\xaa\xe0\xad\x7a\xd0\xf9\xd4\xb6\xad\x3b\x54\x09\
                           \x87\x46\xd4\x52\x4d\x38\x40\x7a\x6d\xeb\x3a\xb7\x8f\xab\x78\xc9";
-        check_keystream(key[], nonce[], keystream);
+        check_keystream(&key, &nonce, keystream);
     }
 }
