+++ tests/rot13.rs
@@ -1,5 +1,5 @@
 #![feature(plugin)]
-#![plugin(nadeko)]
+#![plugin(nadeko_plugin)]
 
 extern crate nadeko;
 
@@ -27,7 +27,6 @@ fn test_rot13() {
     for a in b'A' .. b'Z' + 1 {
         let expected = if a <= b'M' { a + 13 } else { a - 13 };
         let output = rot13::encrypt(a);
-        println!("input: {}", a);
         assert_eq!(expected, output);
     }
 }
